feat: generate filters adapter from registrations, normalize param types
Replace hand-written impl filters::Guest (597 lines) with include! of
build.rs-generated adapter. Build.rs maps Rust types to WIT binding
types (&[f32]→Vec<f32>, &str→String, etc.).

Normalize 3 filter signatures for WIT compatibility:
- bokeh_blur: BokehShape → u32 (0=disc, 1=hexagon)
- colorize: [u8;3] → target_r/target_g/target_b: u8
- convolve: usize → u32 for kw/kh
- perspective_warp: &[f64;9] → &[f64] with length check

Pipeline adapter still needs generation (14 of 123 filter nodes exist).
Full WASM build requires pipeline node generation — tracked separately.

+++ crates/rasmcore-image/src/domain/pipeline/nodes/filters.rs
@@ -337,7 +337,7 @@ impl ImageNode for ConvolveNode {
             &self.source_info,
             self.upstream,
             upstream_fn,
-            |px, info| filters::convolve(px, info, kernel, kw, kh, divisor),
+            |px, info| filters::convolve(px, info, kernel, kw as u32, kh as u32, divisor),
         )
     }
     fn overlap(&self) -> Overlap {
